requirements.txt
opencv-python
mediapipe>=0.10.0
numpy
simpleaudio
PyYAML
pyttsx3